requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
httpx
python-dotenv
passlib[bcrypt]
python-jose
slowapi
email-validator